requirements/polygon_mcp_grpc.txt
# Polygon MCP + gRPC Production System Requirements
# Optimized for polygon_universal_bridge_server.py and MCP layer

# Include base dependencies
-r base.txt

# Web Framework - FastAPI for MCP+gRPC (primary)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Flask for backward compatibility with legacy endpoints
flask>=2.3.3
flask-cors>=4.0.0

# Async and gRPC support
aiohttp>=3.8.0
grpcio>=1.60.0
grpcio-tools>=1.60.0

# Data validation and serialization
pydantic>=2.5.0

# Scientific computing for mathematical optimizations
numpy>=1.24.0

# Performance monitoring
prometheus-client>=0.16.0

# Advanced data structures for MCP layer (included in base.txt)

# Optional: Enhanced performance
# asyncio-throttle>=1.0.0
# aiodns>=3.0.0 